Create scraper.py
The main aim of this python file is to continuously check for the price of a particular object. In this case I wanted to be notified when the price of my lenovo laptop screen drops. I run this code 24/7 on my raspberry pi and will continue to do so until I no longer need to. I made my email and password hidden (for obv reasons) so just input yours if you want to use it.